make player projectile separate from player body + animations adjustments
