+++ Assets/Animations/Player Projectile/Explosion.anim.meta
