+++ Assets/Animations/Player Projectile/Flying.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: PlayerProjectile_Charging
+  m_Name: Flying
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,13 +21,7 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: -403435105, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - time: 0.083333336
-      value: {fileID: -1392887387, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - time: 0.16666667
-      value: {fileID: 2052426297, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - time: 0.25
-      value: {fileID: 2080792067, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 21300000, guid: d190400f966dc22418a34e4a9092bea3, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -50,21 +44,18 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: -403435105, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: -1392887387, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 2052426297, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 2080792067, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 21300000, guid: d190400f966dc22418a34e4a9092bea3, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.26666668
+    m_StopTime: 0.016666668
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 1
+    m_LoopTime: 0
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
