+++ Assets/Animations/Player Projectile/Flying.anim.meta
