+++ Assets/Animations/Player Projectile/Particles.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: PlayerProjectile_Noise
+  m_Name: Particles
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,15 +21,13 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: 193171848, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 1002917925, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.083333336
-      value: {fileID: 1407938236, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 1043033902, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.16666667
-      value: {fileID: -1678554269, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 233233587, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.25
-      value: {fileID: -801174883, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - time: 0.33333334
-      value: {fileID: -2042455874, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: -941806335, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -52,22 +50,21 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: 193171848, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 1407938236, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: -1678554269, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: -801174883, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: -2042455874, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 1002917925, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 1043033902, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 233233587, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: -941806335, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.35000002
+    m_StopTime: 0.26666668
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 0
+    m_LoopTime: 1
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
