+++ Assets/Animations/Player Projectile/Particles.anim.meta
