+++ Assets/Animations/Player Projectile/Recharging.anim.meta
