+++ Assets/Animations/Player Weapon.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6eb779829170a67ce86dee0738aa47b7
-NativeFormatImporter:
+guid: 31c0c5617e93a875eb057655b0f445dd
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
