+++ Assets/Animations/Player Weapon/Charged.anim.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6eb779829170a67ce86dee0738aa47b7
+guid: e97e39d447c31850ca4ef0d9e97f1011
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
