+++ Assets/Animations/Player Weapon/Charging.anim.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6eb779829170a67ce86dee0738aa47b7
+guid: 11a7727953aad869ca6cefcf401933fb
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
