+++ Assets/Animations/Player Weapon/Disapear.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7f42771c299b45f40a5545d71b589d1d
+guid: f1947d53b7c90878d97888ec7d95fe0e
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
