+++ Assets/Animations/Player Weapon/Focusing.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: PlayerProjectile_Recharging
+  m_Name: Focusing
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,15 +21,15 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: -227499425, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: -2139350719, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.083333336
-      value: {fileID: 160514885, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: -1772700558, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.16666667
-      value: {fileID: 1801410343, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 1761074428, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.25
-      value: {fileID: -1973725321, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: 727742634, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     - time: 0.33333334
-      value: {fileID: 436081578, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+      value: {fileID: -1211395479, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -52,11 +52,11 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: -227499425, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 160514885, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 1801410343, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: -1973725321, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
-    - {fileID: 436081578, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: -2139350719, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: -1772700558, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 1761074428, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: 727742634, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
+    - {fileID: -1211395479, guid: 533e3a76c7e07504aa6b0aab7e3a5a3c, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
@@ -67,7 +67,7 @@ AnimationClip:
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 0
+    m_LoopTime: 1
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
