+++ Assets/Animations/Player Weapon/Focusing.anim.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6eb779829170a67ce86dee0738aa47b7
+guid: 62cfe9b23bdbddd2698f252400bc10d3
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
