+++ Assets/Animations/Player Weapon/PlayerWeapon.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 89905d6e58a0fc74fbd353ded5114f59
+guid: 5bb43dc7082c7e44ab4d6ba5aa57c13c
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
