+++ Assets/Scripts/Player/PlayerWeaponAnimations.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa854d0e47fa45ef2a36fb4651a08f73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
